Tratamento de Excessões

+++ src/main/java/com/allos/pomodoro/service/UsuarioService.java
@@ -1,10 +1,12 @@
 package com.allos.pomodoro.service;
 
 import com.allos.pomodoro.dto.UsuarioDTO;
-import com.allos.pomodoro.entity.Usuario;
 import com.allos.pomodoro.mapper.UsuarioMapper;
 import com.allos.pomodoro.repository.UsuarioRepository;
-import com.allos.pomodoro.service.exception.DataIntegrityException;
+
+import com.allos.pomodoro.exception.ObjectNotFoundException;
+import com.allos.pomodoro.exception.ObjectAlreadyExistsException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,24 +21,31 @@ public class UsuarioService {
     @Autowired
     private UsuarioMapper mapper;
 
-    public Optional<UsuarioDTO> buscarUsuarioId(Long id) throws Exception {
+    public Optional<UsuarioDTO> buscarUsuarioId(Long id){
         if(!repository.findById(id).isPresent()){
-            throw new Exception("Usuario não existe.");
+            throw new ObjectNotFoundException("Usuario não existe.");
         }
         return Optional.of(mapper.toDto(repository.findById(id).get()));
     }
 
     public UsuarioDTO salvar(UsuarioDTO dto){
-        Usuario x = mapper.toEntity(dto);
-        Usuario y= repository.save(x);
-       return mapper.toDto(y);
+        if(repository.findByUsuario(dto.getNome()).isPresent() && dto != null){
+            throw new ObjectAlreadyExistsException("Usuario já cadastrado.");
+        }
+        return mapper.toDto(repository.save(mapper.toEntity(dto)));
     }
 
     public UsuarioDTO atualizarUsuario(UsuarioDTO dto) {
+        if(!repository.findById(dto.getId()).isPresent()){
+            throw new ObjectNotFoundException("Id do usuario não existe.");
+        }
         return mapper.toDto(repository.save(mapper.toEntity(dto)));
     }
 
     public void deletaUsuario(Long id){
+        if(!repository.findById(id).isPresent()){
+            throw new ObjectNotFoundException("Usuario não existe.");
+        }
         repository.deleteById(id);
     }
 }
